pyproject.toml
[tool.poetry]
name = "career-crew"
version = "0.1.0"
description = "AI agents that help advance my career."
authors = ["Nerijus Masikonis <nerijus@masikonis.lt>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.12"
langchain = "^0.3.12"
langchain-core = "^0.3.25"
langgraph = "^0.2.59"
python-dotenv = "^1.0.0"
zenrows = "^1.4.0"
langchain-openai = "^0.2.12"
langchain-community = "^0.3.12"
notion-client = "^2.2.1"
nest-asyncio = "^1.6.0"
langsmith = "^0.2.3"
pinecone = "^5.4.2"
fastapi = "^0.115.6"
uvicorn = "^0.34.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pytest-asyncio = "^0.25.0"
pytest-mock = "^3.14.0"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "strict"
asyncio_default_fixture_loop_scope = "function"
pythonpath = "."

[tool.isort]
profile = "black"  # Makes isort compatible with black
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 88
